persistent user positions

+++ mmda-frontend/src/wordcloud/wordcloud.js
@@ -479,6 +479,8 @@ class WordcloudWindow {
     if (this.pressed_node && this.dragging) {
       if (this.hover_node) {
         this.pressed_node.dropAt(this.hover_node);
+      }else{
+        this.pressed_node.drop();
       }
       this.pressed_node.dragging = false;
     }
@@ -673,6 +675,16 @@ class WordcloudWindow {
   }
 
   setupCoordinates(coordinates){
+    //TODO:: check consistency,... only update when new data arrives.
+    var data_consistent = true;
+    for(var word of Object.keys(coordinates)){
+      var el = this.Map.get(word);
+      if(!el){ data_consistent=false; console.log(word); break;}
+      if(! el.matches( coordinates[ word ] )){ data_consistent=false; break;};
+    }
+
+    if(data_consistent) return;
+
     for(var [_,a] of this.Map.entries()) a.delete();
     this.Map = new Map();
     this.coordinates = coordinates;
